+++ packages/sdk/src/provider/injected/models/injected-wallet-api.ts
@@ -12,7 +12,10 @@ import { hasProperties, hasProperty } from 'src/utils/types';
 
 export interface InjectedWalletApi {
     deviceInfo: DeviceInfo;
-    walletInfo: Pick<WalletInfoDTO, 'name' | 'tondns' | 'image' | 'about_url' | 'platforms'>;
+    walletInfo: Pick<
+        WalletInfoDTO,
+        'name' | 'app_name' | 'tondns' | 'image' | 'about_url' | 'platforms'
+    >;
     protocolVersion: number;
     isWalletBrowser: boolean;
     connect(protocolVersion: number, message: ConnectRequest): Promise<ConnectEvent>;
@@ -34,6 +37,7 @@ export function isJSBridgeWithMetadata(value: unknown): value is { tonconnect: I
 
         return hasProperties(value.tonconnect.walletInfo, [
             'name',
+            'app_name',
             'image',
             'about_url',
             'platforms'