+++ packages/sdk/src/resources/fallback-wallets-list.ts
@@ -2,6 +2,7 @@ import { WalletInfoDTO } from 'src/models/wallet/wallet-info';
 
 export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
     {
+        app_name: 'tonkeeper',
         name: 'Tonkeeper',
         image: 'https://tonkeeper.com/assets/tonconnect-icon.png',
         tondns: 'tonkeeper.ton',
@@ -20,6 +21,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['ios', 'android', 'chrome', 'firefox']
     },
     {
+        app_name: 'openmask',
         name: 'OpenMask',
         image: 'https://raw.githubusercontent.com/OpenProduct/openmask-extension/main/public/openmask-logo-288.png',
         about_url: 'https://www.openmask.app/',
@@ -32,6 +34,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['chrome']
     },
     {
+        app_name: 'mytonwallet',
         name: 'MyTonWallet',
         image: 'https://mytonwallet.io/icon-256.png',
         about_url: 'https://mytonwallet.io',
@@ -49,6 +52,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['chrome', 'windows', 'macos', 'linux']
     },
     {
+        app_name: 'tonhub',
         name: 'Tonhub',
         image: 'https://tonhub.com/tonconnect_logo.png',
         about_url: 'https://tonhub.com',
@@ -66,6 +70,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['ios', 'android']
     },
     {
+        app_name: 'tonflow',
         name: 'TonFlow',
         image: 'https://tonflow.net/assets/images/tonflow_ico_192.png',
         about_url: 'https://tonflow.net',
@@ -78,6 +83,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['chrome']
     },
     {
+        app_name: 'dewallet',
         name: 'DeWallet',
         image: 'https://app.delabwallet.com/logo_black.png',
         about_url: 'https://delabwallet.com',
@@ -90,6 +96,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['chrome']
     },
     {
+        app_name: 'xtonwallet',
         name: 'XTONWallet',
         image: 'https://xtonwallet.com/assets/img/icon-256-back.png',
         about_url: 'https://xtonwallet.com',
@@ -102,6 +109,7 @@ export const FALLBACK_WALLETS_LIST: WalletInfoDTO[] = [
         platforms: ['chrome', 'firefox']
     },
     {
+        app_name: 'tonwallet',
         name: 'TON Wallet',
         image: 'https://wallet.ton.org/assets/ui/qr-logo.png',
         about_url: