+++ packages/sdk/src/utils/url.ts
@@ -9,3 +9,8 @@ export function removeUrlLastSlash(url: string): string {
 export function addPathToUrl(url: string, path: string): string {
     return removeUrlLastSlash(url) + '/' + path;
 }
+
+export function isTelegramUrl(link: string): boolean {
+    const url = new URL(link);
+    return url.protocol === 'tg:' || url.hostname === 't.me';
+}